+++ backend/src/churches/members/controller/members.controller.ts
@@ -31,7 +31,6 @@ export class MembersController {
     @Param('churchId', ParseIntPipe) churchId: number,
     @Query() dto: GetMemberDto,
   ) {
-    //console.log(dto);
     return this.membersService.getMembers(churchId, dto);
   }
